requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu
python-multipart
google-generativeai
groq
duckduckgo-search
pytest
httpx
itsdangerous
pydantic[email]
PyMuPDF
beautifulsoup4
langdetect
pycountry
aiosqlite
python-jose[cryptography]
pytesseract
google-cloud-vision
azure-cognitiveservices-speech
authlib
passlib[bcrypt]
openai
deep-translator
python-multipart